+++ R/kappaGold-package.R
@@ -3,7 +3,9 @@
 #' 
 #' Estimate agreement with a gold-standard rating for nominal categories.
 #' 
-#' @importFrom tibble tibble
+#' @importFrom purrr map walk
+#' @importFrom rlang `%||%`
+#' @importFrom tibble tibble lst
 #' @importFrom tidyr pivot_longer pivot_wider
 #' @name kappaGold
 #' @docType package