+++ tests/testthat/test-vanbelle.R
@@ -2,10 +2,11 @@
 
 test_that("kappa vanbelle", {
 
-  # taken from vanbelle, albert paper, see Table 5, first row "proposed method"
+  SCt_idx_gr1 <- which(startsWith(colnames(SC_test), "S"))
+  # taken from Vanbelle & Albert paper, see Table 5, first row "proposed method"
   #+they seem to have used linear weights
-  kvb_sct <- kappam_vanbelle(ratingsGr1 = SC_test[, startsWith(colnames(SC_test), "S")],
-                             ratingsGr2 = SC_test[, startsWith(colnames(SC_test), "E")],
+  kvb_sct <- kappam_vanbelle(ratings = SC_test,
+                             refIdx = SCt_idx_gr1,
                              ratingScale = -2:2,
                              weights = "linear")
   
@@ -15,4 +16,11 @@ test_that("kappa vanbelle", {
                             "se", "conf.level", "ci.lo", "ci.hi", "ci.width"))
   expect_equal(kvb_sct$value, expected = 0.72, tolerance = 1e-2)
   expect_equal(kvb_sct$se, expected = 0.049, tolerance = 1e-2)
+
+  # same results when inverting group1 <=> group2
+  expect_equal(kappam_vanbelle(ratings = SC_test,
+                                   refIdx = -SCt_idx_gr1,
+                                   ratingScale = -2:2,
+                                   weights = "linear"),
+               expected = kvb_sct)
 })